Check if file exists with install

+++ billmategateway/setup/InitInstall.php
@@ -49,12 +49,15 @@
 				$this->db->execute('DELETE FROM '._DB_PREFIX_.'module WHERE name = "billmatepartpay";');
 				$this->db->execute('DELETE FROM '._DB_PREFIX_.'module WHERE name = "billmatecardpay";');
 				$this->db->execute('DELETE FROM '._DB_PREFIX_.'module WHERE name = "billmateinvoice";');
-				$this->db->execute('DELETE FROM '._DB_PREFIX_.'module WHERE name = "billmateinvoiceservice";');
+				if(file_exists(_PS_MODULE_DIR_.'/billmategateway/methods/Invoiceservice.php')) {
+					$this->db->execute('DELETE FROM ' . _DB_PREFIX_ . 'module WHERE name = "billmateinvoiceservice";');
+					$this->db->execute('INSERT INTO '._DB_PREFIX_.'module (name,active,version) VALUES("billmateinvoiceservice",1,"2.0.0");');
+				}
 				$this->db->execute('INSERT INTO '._DB_PREFIX_.'module (name,active,version) VALUES("billmatebankpay",1,"2.0.0");');
 				$this->db->execute('INSERT INTO '._DB_PREFIX_.'module (name,active,version) VALUES("billmateinvoice",1,"2.0.0");');
 				$this->db->execute('INSERT INTO '._DB_PREFIX_.'module (name,active,version) VALUES("billmatecardpay",1,"2.0.0");');
 				$this->db->execute('INSERT INTO '._DB_PREFIX_.'module (name,active,version) VALUES("billmatepartpay",1,"2.0.0");');
-				$this->db->execute('INSERT INTO '._DB_PREFIX_.'module (name,active,version) VALUES("billmateinvoiceservice",1,"2.0.0");');
+
 			}
 			catch (Exception $e)
 			{
